requirements.txt
google-adk>=1.27.2
pydantic>=2.12.5
typing>=3.10.0.0
type-extensions